+++ core/build.gradle.kts
@@ -6,16 +6,13 @@ plugins {
 
 dependencies {
     val springVersion by properties
-    val jjwtVersion by properties
 
     api(project(":port-in"))
     implementation(project(":adapter-out:rdb"))
     implementation(project(":adapter-out:social"))
+    implementation(project(":support:jwt"))
 
     implementation("org.springframework.boot:spring-boot-starter:$springVersion")
-    implementation("io.jsonwebtoken:jjwt-api:$jjwtVersion")
-    runtimeOnly("io.jsonwebtoken:jjwt-impl:$jjwtVersion")
-    runtimeOnly("io.jsonwebtoken:jjwt-jackson:$jjwtVersion")
 
     testImplementation("org.springframework.boot:spring-boot-starter-test:$springVersion")
 }
